+++ cli/src/pcluster/validators/slurm_settings_validator.py
@@ -10,40 +10,86 @@
 # limitations under the License.
 
 from enum import Enum
+from typing import Dict, List
 
 from pcluster.validators.common import FailureLevel, Validator
 
 # SLURM SETTINGS are case-insensitive - keep them lowercase since they are compared with setting.lower()
 SLURM_SETTINGS_DENY_LIST = {
-    "Queue": ["nodes", "partitionname", "resumetimeout", "state", "suspendtime"],
-    "ComputeResource": ["cpus", "features", "gres", "nodeaddr", "nodehostname", "nodename", "state"],
+    "SlurmConf": {
+        "Global": [
+            "communicationparameters",
+            "grestypes",
+            "jobcomppass",
+            "launchparameters",
+            "reconfigflags",
+            "resumefailprogram",
+            "resumeprogram",
+            "resumetimeout",
+            "slurmctldhost",
+            "slurmctldlogfile",
+            "slurmctldparameters",
+            "slurmdlogfile",
+            "slurmuser",
+            "suspendprogram",
+            "suspendtime",
+            "taskplugin",
+            "treewidth",
+        ],
+        "Accounting": [
+            "accountingstoragetype",
+            "accountingstoragehost",
+            "accountingstorageport",
+            "accountingstorageuser",
+            "jobacctgathertype",
+        ],
+    },
+    "Queue": {
+        "Global": ["nodes", "partitionname", "resumetimeout", "state", "suspendtime"],
+    },
+    "ComputeResource": {
+        "Global": ["cpus", "features", "gres", "nodeaddr", "nodehostname", "nodename", "state"],
+    },
 }
 
 
-class SlurmCustomSettingLevel(str, Enum):
+class CustomSlurmSettingLevel(str, Enum):
     """
-    Slurm Custom Settings level.
+    Custom Slurm Settings level.
 
     This enum defines the scope where the custom settings are defined.
     """
 
+    SLURM_CONF = "SlurmConf"
     QUEUE = "Queue"
     COMPUTE_RESOURCE = "ComputeResource"
 
 
-class SlurmCustomSettingsValidator(Validator):
+class CustomSlurmSettingContext(Enum):
     """
-    Slurm Custom Settings validator.
+    Custom Slurm Settings context.
+
+    This enum defines the context where the custom settings are relevant (useful for validation purposes only).
+    """
+
+    GLOBAL = "Global"
+    ACCOUNTING = "Accounting"
+
+
+class CustomSlurmSettingsValidator(Validator):
+    """
+    Custom Slurm Settings validator.
 
     Validate custom settings in Slurm ComputeResource and Queue.
     """
 
-    def _validate(self, custom_settings, deny_list, settings_level: SlurmCustomSettingLevel):
+    def _validate(self, custom_settings: List[Dict], deny_list: List[str], settings_level: CustomSlurmSettingLevel):
         denied_settings = set()
 
-        for custom_setting in list(custom_settings.keys()):
-            if custom_setting.lower() in deny_list:
-                denied_settings.add(custom_setting)
+        for custom_settings_dict in custom_settings:
+            for custom_setting in list(custom_settings_dict.keys()):
+                if custom_setting.lower() in deny_list:
+                    denied_settings.add(custom_setting)
         if len(denied_settings) > 0:
             settings = ",".join(sorted(denied_settings))
             self._add_failure(
@@ -52,9 +98,9 @@ class SlurmCustomSettingsValidator(Validator):
             )
 
 
-class SlurmCustomSettingsWarning(Validator):
+class CustomSlurmSettingsWarning(Validator):
     """
-    Slurm Custom Settings Warning.
+    Custom Slurm Settings Warning.
 
     This validator emits a warning message if custom settings are enabled.
     The message is displayed only once no matter how many times instances of the validator are created.
@@ -63,9 +109,55 @@ class SlurmCustomSettingsWarning(Validator):
     signaled = False
 
     def _validate(self):
-        if not SlurmCustomSettingsWarning.signaled:
+        if not CustomSlurmSettingsWarning.signaled:
             self._add_failure(
                 "Custom Slurm settings are in use: please monitor the cluster carefully.",
                 FailureLevel.WARNING,
             )
-            SlurmCustomSettingsWarning.signaled = True
+            CustomSlurmSettingsWarning.signaled = True
+
+
+class CustomSlurmNodeNamesValidator(Validator):
+    """
+    Custom Slurm Nodelists Names validator.
+
+    This validator ensures that any eventual custom node list passed via SlurmSettings/CustomSlurmSettings
+    does not contain the `-st-` or `-dy-` patterns in the node names, as this would cause the ParallelCluster
+    daemons to interfere with them.
+    """
+
+    def _validate(self, custom_settings: List[Dict]):
+        bad_nodelists = []
+
+        for custom_settings_dict in custom_settings:
+            # Here we validate also the corner case where users provide `NodeName` multiple times with more than
+            # one combination of cases (e.g. `NodeName` and `nodename`)
+            nodenames = [custom_settings_dict[key] for key in custom_settings_dict.keys() if key.lower() == "nodename"]
+            for nodename in nodenames:
+                if ("-st-" in nodename) or ("-dy-" in nodename):
+                    bad_nodelists.append(nodename)
+
+        if bad_nodelists:
+            nodelists = ", ".join(sorted(bad_nodelists))
+            self._add_failure(
+                f"Substrings '-st-' and '-dy-' in node names are reserved for nodes managed by ParallelCluster. "
+                f"Please rename the following custom Slurm nodes: {nodelists}",
+                FailureLevel.ERROR,
+            )
+
+
+class CustomSlurmSettingsIncludeFileOnlyValidator(Validator):
+    """
+    Custom Slurm Settings Include File Only validator.
+
+    This validator returns an error if the CustomSlurmSettingsIncludeFile configuration parameter
+    is used together with the CustomSlurmSettings under SlurmSettings.
+    """
+
+    def _validate(self, custom_settings: List[Dict], include_file_url: str):
+        if custom_settings and include_file_url:
+            self._add_failure(
+                "CustomSlurmsettings and CustomSlurmSettingsIncludeFile cannot be used together "
+                "under SlurmSettings.",
+                FailureLevel.ERROR,
+            )